renamed Command to RCTCommand and moved RemoteCommandInterpreter to claw.rct.remote
